Move all controller related specs into one file
